[r] Remove code handling an impossible combination of state
VersionType.create_only is only used by contributions and contributions are never deleted individually. Contributions are deleted either when an entire index is removed or a with a delete-by-query when deindexing a source. In other words, the combination of `VersionType.create_only` and OpType.delete can not occur when the deleted code is executed.