Remove the envhook hack for sharing cached AWS credentials
c4404942 eliminated the annoying, premature MFA prompt but the warning replacing the prompt was displayed too soon/often. It was safe to remove the hack because the sharing of the credential cache *is* set up when obtaining Boto client instances via azul.deployment.aws, which covers all required use cases. Stock Boto3 clients won't share cached credentials, and neither will `terraform` processes launched from a Python process that is a child of PyCharm but not a child of a bash process where _login_aws was invoked. Neither of these cases apply to day-to-day development or use of Azul.